+++ CHANGELOG.md
@@ -1,24 +1,32 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## All [unreleased]
+
 ### Added
+
 - Print connected user when getting loans
 - Display due dates as date, not datetime
 
 ## [0.2.0] - 2022-10-25
+
 ### Added
+
 - Ability to get the current loans of a user
 
 ### Changed
+
 - Drop support of Python 3.7 - 3.9. Only support Python 3.10
 
 ## [0.1.0] - 2022-10-05
+
 ### Added
+
 - Initial release
 
-[Unreleased]: https://github.com/tomsquest/mediathequeroubaix.py/compare/0.2.0...master
+[unreleased]: https://github.com/tomsquest/mediathequeroubaix.py/compare/0.2.0...master
 [0.2.0]: https://github.com/tomsquest/mediathequeroubaix.py/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/tomsquest/mediathequeroubaix.py/tree/0.1.0