+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2022-11-22
 ### Added
 - Add CLI (command-line interface) to create a sample config, show the config and get loans
 - Display loans in a pretty table
@@ -21,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/tomsquest/mediathequeroubaix.py/compare/0.2.0...master
+[Unreleased]: https://github.com/tomsquest/mediathequeroubaix.py/compare/1.0.0...master
+[1.0.0]: https://github.com/tomsquest/mediathequeroubaix.py/compare/0.2.0...1.0.0
 [0.2.0]: https://github.com/tomsquest/mediathequeroubaix.py/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/tomsquest/mediathequeroubaix.py/tree/0.1.0